requirements.txt
streamlit
numpy
Pillow
opencv-python-headless
PyYAML
tensorflow
